requirements.txt
fastapi==0.77.1
SQLAlchemy==1.4.36
python-dotenv==0.20.0
uvicorn==0.17.6
psycopg2-binary==2.9.3
